requirements.txt
Django==3.2.16
Pillow==9.3.0
asgiref==3.6.0
pip==21.3.1
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.1
gunicorn==20.1.0
whitenoise==6.2.0
django-dynamic-admin-forms==0.1.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
colorama==0.4.6
django-mathfilters==1.0.0
loguru==0.6.0
pytz==2022.7
dj-database-url==1.2.0